+++ photutils/background/tests/test_background_2d.py
@@ -140,17 +140,7 @@ class TestBackground2D:
 
         assert bkg2.background_mesh[0, 0] > bkg1.background_mesh[0, 0]
 
-    @pytest.mark.parametrize('filter_size', FILTER_SIZES)
-    def test_resizing(self, filter_size):
-        with pytest.warns(AstropyDeprecationWarning):
-            bkg1 = Background2D(DATA, (23, 22), filter_size=filter_size,
-                                bkg_estimator=MeanBackground(),
-                                edge_method='crop')
-        bkg2 = Background2D(DATA, (23, 22), filter_size=filter_size,
-                            bkg_estimator=MeanBackground())
-        assert_allclose(bkg1.background, bkg2.background, rtol=2e-6)
-        assert_allclose(bkg1.background_rms, bkg2.background_rms)
-
+    def test_resizing(self):
         shape1 = (128, 256)
         shape2 = (129, 256)
         box_size = (16, 16)
@@ -179,13 +169,6 @@ class TestBackground2D:
                            bkg_estimator=MeanBackground())
         assert_allclose(bkg.background, DATA, rtol=2.0e-5)
         assert_allclose(bkg.background_rms, BKG_RMS)
-
-        # test edge crop with mask
-        with pytest.warns(AstropyDeprecationWarning):
-            bkg2 = Background2D(data, box_size, filter_size=(1, 1), mask=mask,
-                                bkg_estimator=MeanBackground(),
-                                edge_method='crop')
-        assert_allclose(bkg2.background, DATA, rtol=2.0e-5)
 
     def test_mask(self):
         data = np.copy(DATA)
@@ -445,13 +428,6 @@ class TestBackground2D:
             Background2D(DATA, (25, 25), filter_size=(1, 1),
                          coverage_mask=np.zeros((2, 2, 2)))
 
-    def test_invalid_edge_method(self):
-        match = 'edge_method must be "pad" or "crop"'
-        with (pytest.warns(AstropyDeprecationWarning),
-              pytest.raises(ValueError, match=match)):
-            Background2D(DATA, (23, 22), filter_size=(1, 1),
-                         edge_method='not_valid')
-
     @pytest.mark.skipif(not HAS_MATPLOTLIB, reason='matplotlib is required')
     def test_plot_meshes(self):
         """
@@ -460,14 +436,6 @@ class TestBackground2D:
         """
         bkg = Background2D(DATA, (25, 25))
         bkg.plot_meshes(outlines=True)
-
-    def test_crop(self):
-        data = np.ones((300, 500))
-        with pytest.warns(AstropyDeprecationWarning):
-            bkg = Background2D(data, (74, 99), edge_method='crop')
-        assert_allclose(bkg.background_median, 1.0)
-        assert_allclose(bkg.background_rms_median, 0.0)
-        assert_allclose(bkg.background_mesh.shape, (4, 5))
 
     def test_repr(self):
         data = np.ones((300, 500))